# frontend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "News API",
    "description": "HTTP API for browsing aggregated news articles and managing user access.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Service health check",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Service is reachable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/news": {
      "get": {
        "summary": "List news",
        "operationId": "listNews",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by exact category name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period",
            "in": "query",
            "description": "Filter by relative publish period",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/NewsPeriod"
            }
          },
          {
            "name": "sources",
            "in": "query",
            "description": "Comma separated list of sources to include",
            "required": false,
            "schema": {
              "type": "string"
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of articles to return",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of news items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NewsItem"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid filter parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/news/{news_id}": {
      "get": {
        "summary": "News item details",
        "operationId": "newsDetail",
        "parameters": [
          {
            "name": "news_id",
            "in": "path",
            "required": true,
            "description": "Unique news identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full news article data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetail"
                }
              }
            }
          },
          "404": {
            "description": "News item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sources": {
      "get": {
        "summary": "List available sources",
        "operationId": "listSources",
        "responses": {
          "200": {
            "description": "Available RSS feeds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register new user",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid registration data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "User login",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid credentials payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          }
        }
      },
      "NewsItem": {
        "type": "object",
        "required": [
          "id",
          "title",
          "summary",
          "category",
          "source"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "link": {
            "type": "string",
            "format": "uri",
            "nullable": true
          }
        }
      },
      "NewsDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/NewsItem"
          },
          {
            "type": "object",
            "properties": {
              "content": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "SourceInfo": {
        "type": "object",
        "required": [
          "name",
          "chosen",
          "available"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "chosen": {
            "type": "boolean",
            "description": "True when the source is selected by default"
          },
          "available": {
            "type": "boolean",
            "description": "True when recent articles from the source are present in storage"
          }
        }
      },
      "Register": {
        "type": "object",
        "required": [
          "name",
          "password",
          "preferences"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "preferences": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Login": {
        "type": "object",
        "required": [
          "name",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          }
        }
      },
      "NewsPeriod": {
        "type": "string",
        "enum": [
          "day",
          "week",
          "month"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "description": "Error details",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            ]
          }
        }
      }
    }
  }
}